backend/ai-python/requirements.txt
fastapi
uvicorn
ultralytics
opencv-python-headless
torch
torchvision
numpy
pillow
python-multipart
google-generativeai
python-dotenv